[YottaDB/DB/YDB#1007] Bubble up XCRETNULLREF
Any returns from the API due to error conditions where a NULL is
returned will raise a trappable M error, not just go in the system log.

Previous commit 6393e5d14c22931970aca99d16b866cd08ad2b8d notes that
users should ignore XCRETNULLREF if it shows up in the error log as
YottaDB ignored that error and just put it in the log. This is now no
longer true, and is the better behavior, as all of these messages imply
some sort of error condition.